requirements.txt
pandas==2.0.3
Pillow==10.0.0
plotly==5.16.1
streamlit==1.25.0
